Move useDebouncedEffect to a separate file.
